// backend/src/endpoints/company/dto/update-factory.dto.ts
//
// Copyright (c) 2026 IndustryFusion Europe UG
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
// http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
//

export interface UpdateFactoryDto {
  location_name?: string;
  address_1?: string;
  city?: string;
  country?: string;
  zip?: string;
  latitude?: number;
  longitude?: number;
  timezone?: string;
}